Fixes #53 (partially)
This change simplifies the entry point of the React application by
moving the root component code into its own file and importing.
